requirements.txt
streamlit
openai
langchain_anthropic
anthropic
audio-recorder-streamlit
lngchain
python-decouple
moviepy